user edit and delete

+++ app/models/latest.rb
@@ -16,7 +16,7 @@ class Latestproduct
 	validates_attachment_content_type :image, :content_type => ["image/jpg", "image/jpeg", "image/png", "image/gif"]
 	field :rating, :type => Float
 	field :views, :type => Integer, :default => 0
-	field :avg_rating, :type => Integer, :default => -1
+	field :avg_rating, :type => Float, :default => -1
 	embedded_in :latest, :inverse_of => :latestproduct
 	search_in :product_name
 end
